+++ R/cnd.R
@@ -682,6 +682,14 @@ conditionMessage.rlang_error <- function(c) {
   paste_line(message, backtrace)
 }
 
+#' @export
+as.character.rlang_error <- function(x, ...) {
+  # Don't generate backtrace or reminder in programmatic uses. Fixes
+  # backtraces in knitr.
+  x$message
+}
+
+
 #' Display backtrace on error
 #'
 #' @description