+++ tests/testthat/helper-trace.R
@@ -21,3 +21,20 @@ expect_equal_trace <- function(x, y) {
   expect_identical(x$parents, y$parents)
   expect_equal(x$calls, y$calls)
 }
+
+render_md <- function(file) {
+  skip_if_not_installed("rmarkdown")
+
+  out_file <- tempfile(file)
+  on.exit(file.remove(out_file))
+
+  rmarkdown::render(
+    test_path(file),
+    output_format = "md_document",
+    output_file = out_file,
+    quiet = TRUE,
+    envir = globalenv()
+  )
+
+  readLines(out_file)
+}